core+ui: movable border-band seams — fractional positions + constrained drag (B5.5)
Border span endpoints (start_node/end_node) generalize from node indices to
fractional boundary positions (schema integer -> double; old plans still load),
so a band can start or end mid-edge. The offset-ribbon renderer and interval
offset-stacking are fully fractional, and cost via boundary_span_length walks an
arc-length coordinate with rem_euclid wrap.

Each span endpoint gets a draggable seam handle, floated off the paver boundary
onto its band's inner edge (along the corner bisector at a node) so the dots
form an inward array clear of the area's edge nodes. Dragging is constrained by
slp-core::constrain_seam: confined to the seam's associated edge (projected
along the true curve for arc/Bezier edges, gap measured along the arc), snapped
to a free end within 1 ft, and held a 1.5 ft gap off the band's own other
endpoint (per-border, so different borders on an edge may align) — so seams
can't leave their edge, cross, or pile up at a corner.

The border-row editor is now two stacked lines (material + width + remove, then
From/To), so the material and node numbers stay readable.

arc.rs is mutation-clean (0 missed); seam_edge, constrain_seam, and edge_point
are unit-tested for straight/arc/Bezier edges.

+++ crates/slp-ui/src/components/area_inspector.rs
@@ -118,11 +118,11 @@ pub fn AreaInspector(
     #[prop(default = 0)]
     border_node_count: usize,
     /// Set ring `i`'s span start node (`None` = whole perimeter).
-    #[prop(default = Callback::new(|_: (usize, Option<i64>)| {}))]
-    on_border_start: Callback<(usize, Option<i64>)>,
+    #[prop(default = Callback::new(|_: (usize, Option<f64>)| {}))]
+    on_border_start: Callback<(usize, Option<f64>)>,
     /// Set ring `i`'s span end node (`None` = whole perimeter).
-    #[prop(default = Callback::new(|_: (usize, Option<i64>)| {}))]
-    on_border_end: Callback<(usize, Option<i64>)>,
+    #[prop(default = Callback::new(|_: (usize, Option<f64>)| {}))]
+    on_border_end: Callback<(usize, Option<f64>)>,
     /// Append a border ring.
     #[prop(default = Callback::new(|()| {}))]
     on_border_add: Callback<()>,
